data_generator/requirements.txt
langchain
langchain-community
langchain-openai
langchain-anthropic
langchain-google-genai
langchain-ollama
openai
anthropic
google-generativeai
ollama
PyPDF2
pdfplumber
markdown
python-docx
unstructured
pydantic>=2.0.0
python-dotenv
requests
tqdm

